fix: extraced the useUpdateMinZoom hooks from the DynamicMinZoom component
